+++ docs/changes.rst
@@ -6,6 +6,9 @@ Version 0.3.0
 
 To be released.
 
+- Added :attr:`Image.metadata <wand.image.Image.metadata>` immutable mapping
+  attribute and :class:`~wand.image.Metadata` mapping type for it.
+
 
 Version 0.2.2
 -------------